# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LSU Football player database API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface Player {
  playerId: string;
  playerName: string;
  /** @nullable */
  team?: string | null;
  /** @nullable */
  position?: string | null;
  /** @nullable */
  posGroup?: string | null;
  /** @nullable */
  conference?: string | null;
  /** @nullable */
  jersey?: string | null;
  season: number;
  /** @nullable */
  week?: string | null;
  /** @nullable */
  snapsNonSt?: number | null;
  /** @nullable */
  snapsSt?: number | null;
  /** @nullable */
  war?: number | null;
  /** @nullable */
  twar?: number | null;
  /** @nullable */
  par?: number | null;
  /** @nullable */
  playerValue?: number | null;
  /** @nullable */
  playerValuePct?: number | null;
  /** @nullable */
  playerTier?: string | null;
}

export interface GradeItem {
  key: string;
  label: string;
  value: number;
  /** @nullable */
  category?: string | null;
}

export interface PlayerDetail {
  playerId: string;
  playerName: string;
  /** @nullable */
  team?: string | null;
  /** @nullable */
  position?: string | null;
  /** @nullable */
  posGroup?: string | null;
  /** @nullable */
  conference?: string | null;
  /** @nullable */
  jersey?: string | null;
  season: number;
  /** @nullable */
  week?: string | null;
  /** @nullable */
  snapsNonSt?: number | null;
  /** @nullable */
  snapsSt?: number | null;
  /** @nullable */
  war?: number | null;
  /** @nullable */
  twar?: number | null;
  /** @nullable */
  par?: number | null;
  /** @nullable */
  playerValue?: number | null;
  /** @nullable */
  playerValuePct?: number | null;
  /** @nullable */
  playerTier?: string | null;
  grades?: GradeItem[];
}

export interface PlayerListResponse {
  players: Player[];
  total: number;
  page: number;
  pageSize: number;
}

export interface Team {
  school: string;
  /** @nullable */
  mascot?: string | null;
  /** @nullable */
  abbreviation?: string | null;
  /** @nullable */
  conference?: string | null;
  /** @nullable */
  classification?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  altColor?: string | null;
  /** @nullable */
  logo?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  playerCount?: number | null;
}

export interface TeamDetail {
  team: Team;
  roster: Player[];
}

export interface PositionGroupStat {
  posGroup: string;
  count: number;
  /** @nullable */
  avgWar?: number | null;
  /** @nullable */
  avgValue?: number | null;
  /** @nullable */
  topPlayerName?: string | null;
  /** @nullable */
  topPlayerWar?: number | null;
}

export interface DashboardSummary {
  totalPlayers: number;
  totalTeams: number;
  /** @nullable */
  avgWar?: number | null;
  topPlayer?: Player;
  seasons: number[];
  positionGroups?: PositionGroupStat[];
  topByValue?: Player[];
}

export interface FilterOptions {
  seasons: number[];
  teams: string[];
  conferences: string[];
  posGroups: string[];
  positions: string[];
}

export interface SourceStatus {
  name: string;
  configured: boolean;
  /** @nullable */
  ok?: boolean | null;
  /** @nullable */
  detail?: string | null;
}

export interface SyncStatus {
  status: string;
  /** @nullable */
  lastSyncAt?: string | null;
  /** @nullable */
  playersSynced?: number | null;
  /** @nullable */
  teamsSynced?: number | null;
  /** @nullable */
  season?: number | null;
  /** @nullable */
  message?: string | null;
  sources?: SourceStatus[];
}

export interface SyncInput {
  season?: number;
  team?: string;
  conference?: string;
}

export type ListPlayersParams = {
search?: string;
team?: string;
conference?: string;
posGroup?: string;
position?: string;
season?: number;
sort?: ListPlayersSort;
order?: ListPlayersOrder;
page?: number;
pageSize?: number;
};

export type ListPlayersSort = typeof ListPlayersSort[keyof typeof ListPlayersSort];


export const ListPlayersSort = {
  war: 'war',
  twar: 'twar',
  player_value: 'player_value',
  name: 'name',
  snaps: 'snaps',
} as const;

export type ListPlayersOrder = typeof ListPlayersOrder[keyof typeof ListPlayersOrder];


export const ListPlayersOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;

export type ListTeamsParams = {
conference?: string;
};

export type GetDashboardSummaryParams = {
season?: number;
team?: string;
};

export type GetTopPlayersParams = {
season?: number;
team?: string;
posGroup?: string;
metric?: GetTopPlayersMetric;
limit?: number;
};

export type GetTopPlayersMetric = typeof GetTopPlayersMetric[keyof typeof GetTopPlayersMetric];


export const GetTopPlayersMetric = {
  war: 'war',
  twar: 'twar',
  player_value: 'player_value',
} as const;

export type GetPositionGroupsParams = {
season?: number;
team?: string;
};