# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LSU Football player database API
 * OpenAPI spec version: 0.1.0
 */
import type { Player } from './player';
import type { PositionGroupStat } from './positionGroupStat';

export interface DashboardSummary {
  totalPlayers: number;
  totalTeams: number;
  /** @nullable */
  avgWar?: number | null;
  topPlayer?: Player;
  seasons: number[];
  positionGroups?: PositionGroupStat[];
  topByValue?: Player[];
}