# lib/api-zod/src/generated/types/filterOptions.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LSU Football player database API
 * OpenAPI spec version: 0.1.0
 */

export interface FilterOptions {
  seasons: number[];
  teams: string[];
  conferences: string[];
  posGroups: string[];
  positions: string[];
}