# lib/api-zod/src/generated/types/gradeItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LSU Football player database API
 * OpenAPI spec version: 0.1.0
 */

export interface GradeItem {
  key: string;
  label: string;
  value: number;
  /** @nullable */
  category?: string | null;
}