# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LSU Football player database API
 * OpenAPI spec version: 0.1.0
 */

export interface Player {
  playerId: string;
  playerName: string;
  /** @nullable */
  team?: string | null;
  /** @nullable */
  position?: string | null;
  /** @nullable */
  posGroup?: string | null;
  /** @nullable */
  conference?: string | null;
  /** @nullable */
  jersey?: string | null;
  season: number;
  /** @nullable */
  week?: string | null;
  /** @nullable */
  snapsNonSt?: number | null;
  /** @nullable */
  snapsSt?: number | null;
  /** @nullable */
  war?: number | null;
  /** @nullable */
  twar?: number | null;
  /** @nullable */
  par?: number | null;
  /** @nullable */
  playerValue?: number | null;
  /** @nullable */
  playerValuePct?: number | null;
  /** @nullable */
  playerTier?: string | null;
}