# lib/api-zod/src/generated/types/positionGroupStat.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LSU Football player database API
 * OpenAPI spec version: 0.1.0
 */

export interface PositionGroupStat {
  posGroup: string;
  count: number;
  /** @nullable */
  avgWar?: number | null;
  /** @nullable */
  avgValue?: number | null;
  /** @nullable */
  topPlayerName?: string | null;
  /** @nullable */
  topPlayerWar?: number | null;
}