# lib/api-zod/src/generated/types/team.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LSU Football player database API
 * OpenAPI spec version: 0.1.0
 */

export interface Team {
  school: string;
  /** @nullable */
  mascot?: string | null;
  /** @nullable */
  abbreviation?: string | null;
  /** @nullable */
  conference?: string | null;
  /** @nullable */
  classification?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  altColor?: string | null;
  /** @nullable */
  logo?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  playerCount?: number | null;
}