# lib/api-zod/src/generated/types/teamDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LSU Football player database API
 * OpenAPI spec version: 0.1.0
 */
import type { Player } from './player';
import type { Team } from './team';

export interface TeamDetail {
  team: Team;
  roster: Player[];
}